+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.34] - 2026.09.03
 
 ### Added
 
@@ -343,7 +343,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GitLab CI and GitHub Actions deployment examples
 - Agent TOML configuration example
 
-[Unreleased]: https://github.com/studiometa/trafic/compare/0.1.33...HEAD
+[Unreleased]: https://github.com/studiometa/trafic/compare/0.1.34...HEAD
+[0.1.34]: https://github.com/studiometa/trafic/compare/0.1.33...0.1.34
 [0.1.33]: https://github.com/studiometa/trafic/compare/0.1.32...0.1.33
 [0.1.32]: https://github.com/studiometa/trafic/compare/0.1.31...0.1.32
 [0.1.31]: https://github.com/studiometa/trafic/compare/0.1.30...0.1.31